refactor: more fixes after review
