+++ src/main/java/ru/yandex/practicum/filmorate/repository/FilmStorage.java
@@ -1,5 +1,6 @@
 package ru.yandex.practicum.filmorate.repository;
 
+import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.model.Film;
 
 import java.util.ArrayList;
@@ -8,9 +9,9 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+@Component
 public class FilmStorage {
     private final Map<Integer, Film> films = new HashMap<>();
-
 
     public List<Film> getAllFilms() {
         return new ArrayList<>(films.values());
