+++ src/main/java/ru/yandex/practicum/filmorate/repository/UserStorage.java
@@ -1,5 +1,6 @@
 package ru.yandex.practicum.filmorate.repository;
 
+import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.ArrayList;
@@ -8,6 +9,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+@Component
 public class UserStorage {
     private final Map<Integer, User> users = new HashMap<>();
